+++ widgetssdk/src/main/java/com/glia/widgets/base/BaseView.kt
@@ -1,5 +1,5 @@
 package com.glia.widgets.base
 
-internal interface BaseView<T> {
+internal interface BaseView<T : BaseController> {
     fun setController(controller: T)
 }
