+++ widgetssdk/src/main/java/com/glia/widgets/call/CallContract.kt
@@ -1,0 +1,62 @@
+package com.glia.widgets.call
+
+import com.glia.androidsdk.Engagement
+import com.glia.androidsdk.comms.MediaUpgradeOffer
+import com.glia.androidsdk.screensharing.ScreenSharing
+import com.glia.widgets.base.BaseController
+import com.glia.widgets.base.BaseView
+import com.glia.widgets.core.dialog.model.ConfirmationDialogLinks
+import com.glia.widgets.core.dialog.model.Link
+
+internal interface CallContract {
+    interface Controller : BaseController {
+        val confirmationDialogLinks: ConfirmationDialogLinks
+
+        fun onBackClicked()
+        fun leaveChatClicked()
+        fun leaveChatQueueClicked()
+        fun chatButtonClicked()
+        fun onSpeakerButtonPressed()
+        fun minimizeButtonClicked()
+        fun muteButtonClicked()
+        fun videoButtonClicked()
+        fun startCall(
+            companyName: String,
+            queueId: String?,
+            visitorContextAssetId: String?,
+            mediaType: Engagement.MediaType,
+            useOverlays: Boolean,
+            screenSharingMode: ScreenSharing.Mode,
+            upgradeToCall: Boolean
+        )
+
+        fun onResume()
+        fun onPause()
+        fun setView(view: View)
+        fun onLiveObservationDialogRequested()
+        fun endEngagementDialogYesClicked()
+        fun endEngagementDialogDismissed()
+        fun notificationsDialogDismissed()
+        fun acceptUpgradeOfferClicked(mediaUpgradeOffer: MediaUpgradeOffer)
+        fun declineUpgradeOfferClicked(mediaUpgradeOffer: MediaUpgradeOffer)
+        fun noMoreOperatorsAvailableDismissed()
+        fun unexpectedErrorDialogDismissed()
+        fun onLiveObservationDialogAllowed()
+        fun onLiveObservationDialogRejected()
+        fun onLinkClicked(link: Link)
+        fun overlayPermissionsDialogDismissed()
+        fun onUserInteraction()
+        fun shouldShowMediaEngagementView(upgradeToCall: Boolean): Boolean
+        fun onDestroy(retained: Boolean)
+    }
+
+    interface View : BaseView<Controller> {
+        fun emitState(callState: CallState)
+        fun navigateToChat()
+        fun destroyView()
+        fun minimizeView()
+        fun showMissingPermissionsDialog()
+        fun showEngagementConfirmationDialog()
+        fun navigateToWebBrowserActivity(title: String, url: String)
+    }
+}
