+++ widgetssdk/src/main/java/com/glia/widgets/callvisualizer/CallVisualizerSupportActivity.kt
@@ -2,12 +2,12 @@ package com.glia.widgets.callvisualizer
 
 import android.os.Parcelable
 import androidx.appcompat.app.AppCompatActivity
-import com.glia.widgets.callvisualizer.controller.CallVisualizerController
+import com.glia.widgets.callvisualizer.controller.CallVisualizerContract
 import kotlinx.parcelize.Parcelize
 
 class CallVisualizerSupportActivity : AppCompatActivity() {
     companion object {
-        val PERMISSION_TYPE_TAG: String = CallVisualizerController::class.java.simpleName
+        val PERMISSION_TYPE_TAG: String = CallVisualizerContract.Controller::class.java.simpleName
     }
 }
