+++ widgetssdk/src/main/java/com/glia/widgets/callvisualizer/EndScreenSharingView.kt
@@ -6,7 +6,7 @@ import androidx.constraintlayout.widget.ConstraintLayout
 import androidx.core.content.withStyledAttributes
 import com.glia.widgets.R
 import com.glia.widgets.UiTheme
-import com.glia.widgets.core.screensharing.ScreenSharingController
+import com.glia.widgets.core.screensharing.ScreenSharingContract
 import com.glia.widgets.databinding.EndScreenSharingViewBinding
 import com.glia.widgets.di.Dependencies
 import com.glia.widgets.helper.Logger
@@ -43,7 +43,7 @@ internal class EndScreenSharingView(
     var onFinishListener: OnFinishListener? = null
     private var controller: EndScreenSharingContract.Controller? = null
     private var statusBarColor: Int by Delegates.notNull()
-    private var screenSharingController: ScreenSharingController? = null
+    private var screenSharingController: ScreenSharingContract.Controller? = null
     private var defaultStatusBarColor: Int? = null
     private var stringProvider = Dependencies.getStringProvider()
