--- widgetssdk/src/main/java/com/glia/widgets/chat/ChatViewCallback.java
@@ -1,63 +1,0 @@
-package com.glia.widgets.chat;
-
-import android.net.Uri;
-import android.view.View;
-
-import androidx.annotation.NonNull;
-
-import com.glia.androidsdk.chat.AttachmentFile;
-import com.glia.androidsdk.engagement.Survey;
-import com.glia.widgets.chat.model.ChatItem;
-import com.glia.widgets.chat.model.ChatState;
-import com.glia.widgets.core.fileupload.model.FileAttachment;
-
-import org.jetbrains.annotations.NotNull;
-
-import java.util.List;
-
-public interface ChatViewCallback {
-
-    void emitUploadAttachments(List<FileAttachment> attachments);
-
-    void emitState(ChatState chatState);
-
-    void emitItems(List<ChatItem> items);
-
-    void navigateToCall(String requestedMediaType);
-
-    void backToCall();
-
-    void navigateToSurvey(@NonNull Survey survey);
-
-    void destroyView();
-
-    void minimizeView();
-
-    void smoothScrollToBottom();
-
-    void scrollToBottomImmediate();
-
-    void fileDownloadError(AttachmentFile attachmentFile, Throwable error);
-
-    void fileDownloadSuccess(AttachmentFile attachmentFile);
-
-    void clearMessageInput();
-
-    void navigateToPreview(AttachmentFile attachmentFile, View view);
-
-    void fileIsNotReadyForPreview();
-
-    void showBroadcastNotSupportedToast();
-
-    void requestOpenUri(@NonNull Uri uri);
-
-    void requestOpenDialer(@NotNull Uri uri);
-
-    void requestOpenEmailClient(@NotNull Uri uri);
-
-    void showEngagementConfirmationDialog();
-
-    void navigateToWebBrowserActivity(String title, String url);
-
-    void clearTempFile();
-}
